+++ src/phl_risk/data_prep/__init__.py
@@ -5,5 +5,21 @@ from ._base import BasePrepStep
 from ._prep import DataPrep
 from ._result import PrepResult, StepResult
 from ._snapshot import PrepSnapshot
+from .integrations import OptBinningStep
+from .steps import KBinsStep, MissingImputer, SklearnStep, ToDatetime, ToNumeric, ValueMapper
 
-__all__ = ["PrepAudit", "BasePrepStep", "DataPrep", "PrepResult", "StepResult", "PrepSnapshot"]
+__all__ = [
+    "PrepAudit",
+    "BasePrepStep",
+    "DataPrep",
+    "PrepResult",
+    "StepResult",
+    "PrepSnapshot",
+    "OptBinningStep",
+    "KBinsStep",
+    "MissingImputer",
+    "SklearnStep",
+    "ToDatetime",
+    "ToNumeric",
+    "ValueMapper",
+]